fix(duckdb): ensure that casting to floating point values produces valid types in generated sql

+++ ibis/backends/tests/tpch/snapshots/test_h06/test_tpc_h06/duckdb/h06.sql
@@ -4,5 +4,5 @@ FROM main.lineitem AS t0
 WHERE
   t0.l_shipdate >= MAKE_DATE(1994, 1, 1)
   AND t0.l_shipdate < MAKE_DATE(1995, 1, 1)
-  AND t0.l_discount BETWEEN CAST(0.05 AS REAL(53)) AND CAST(0.07 AS REAL(53))
+  AND t0.l_discount BETWEEN CAST(0.05 AS DOUBLE) AND CAST(0.07 AS DOUBLE)
   AND t0.l_quantity < CAST(24 AS TINYINT)
